--- src/components/Footer.js
@@ -1,10 +1,0 @@
-const Footer = () => {
-
-    return (
-        <footer>
-            <p>Created by <a href="https://meiningcheng.com/" target="_blank" rel="noreferrer">Meining Cheng</a> @ Juno College 2021</p>
-        </footer>
-    )
-}
-
-export default Footer;
