+++ src/components/RecipeCard.js
@@ -1,24 +1,15 @@
-
-import { Link } from 'react-router-dom';
-// importing our firebase configuration
-import firebase from "../config/firebase.js";
-
+import { useState } from 'react';
+import RecipeDetails from './RecipeDetails.js';
 import notFoundPhoto from '../assets/notfound_placeholder.svg';
 
-const RecipeCard = ({ recipeData, savedRecipes }) => {
+const RecipeCard = ({ recipeData, addRecipeFunction, savedRecipes, removeRecipe }) => {
 
-    const { foodName, foodImg, calories, recipeSource, key, totalNutrients, totalDaily, totalWeight } = recipeData;
+    const { foodName, foodImg, calories, ingredientList, recipeSource, key } = recipeData;
 
-    // Referencing firebase database
-    const dbRef = firebase.database().ref();
-    const handleAddRecipe = (recipeKey) => {
-        // console.log(recipeKey);
-        dbRef.push(recipeKey);
-    }
+    const [show, setShow] = useState(false);
 
-    const handleRemoveRecipe = (recipeKey) => {
-        // console.log(recipeKey);
-        dbRef.child(recipeKey).remove();
+    const handleShow = () => {
+        setShow(!show);
     }
 
     //check if specific recipe exsit in the list
@@ -38,39 +29,36 @@ const RecipeCard = ({ recipeData, savedRecipes }) => {
     // }).includes(key);
 
     return (
-        // <Router>
-        <li className="recipeListItem">
+        <div className="subContainer">
+
             {
                 isSaved
                     ?
-                    <button className="saveControlButton iconButton" onClick={() => { handleRemoveRecipe(savedRecipe.key) }}>
+                    <button className="heartButton" onClick={() => { removeRecipe(savedRecipe.key) }}>
                         <i className="fas fa-heart" aria-hidden="true"></i>
                         <span className="srOnly">Click to remove recipe remove saved recipe list</span>
                     </button>
                     :
-                    <button className="saveControlButton iconButton" onClick={() => { handleAddRecipe(recipeData) }}>
+                    <button className="heartButton" onClick={() => { addRecipeFunction(recipeData) }}>
                         <i className="far fa-heart" aria-hidden="true"></i>
                         <span className="srOnly">Click to add recipe to saved recipe list</span>
                     </button>
             }
-            <h2 className="recipeHeading">{foodName}</h2>
-            <img src={foodImg} alt={`${foodName}`} onError={(event) => { event.target.src = notFoundPhoto; event.target.alt = "photo not found" }} />
-            <p>{Math.round(calories)} Cal per {Math.round(totalWeight)} g</p>
 
-            <Link to={{
-                pathname: `/recipeDetails/${key}`,
-                state: {
-                    totalNutrients: totalNutrients,
-                    totalDaily: totalDaily,
-                    totalWeight: totalWeight,
-                }
-            }}>
-                <button>Nutrition Facts</button>
-            </Link>
+            <h2>{foodName}</h2>
+
+            <img src={foodImg} alt={`${foodName}`} onError={(event) => { event.target.src = notFoundPhoto; event.target.alt = "photo not found" }} />
+            <p>{parseInt(calories)} Cal</p>
+
+            <button onClick={handleShow}>Ingredient List</button>
+
+            {show ? <RecipeDetails ingredientData={ingredientList} /> : ""}
 
             <a href={recipeSource} target="_blank" rel="noreferrer">Recipe link</a>
 
-        </li>
+            <button>Nutrition Facts</button>
+
+        </div>
     )
 }
