+++ src/components/SavedRecipeContainer.js
@@ -1,45 +1,36 @@
+import { useState } from 'react';
 
-import firebase from "../config/firebase.js";
+const SavedRecipeContainer = ({ recipeListData, removeRecipeFunction }) => {
 
-import notFoundPhoto from '../assets/notfound_placeholder.svg';
-
-const SavedRecipeContainer = ({savedRecipes}) => {
-
-    const dbRef = firebase.database().ref();
-    
-    const handleRemoveRecipe = (recipeKey) => {
-        // console.log(recipeKey);
-        dbRef.child(recipeKey).remove();
-    }
+    const [showList, setShowList] = useState(false);
 
     return (
         <>
-            <div className="recipeContainer wrapper">
-                <h3>Your Recipe List</h3>
-                <p>{savedRecipes.length} recipe{savedRecipes.length > 1 ? "s" : ""}</p>
-                <ul className="savedRecipeList">
-                    {
-                        savedRecipes.map((recipe) => {
-                            // const {key, foodImg} = recipe;
-                            console.log(recipe)
+            <div className="savedListContainer">
+                <button className="heartButton fixedHeart" onClick={() => { setShowList(!showList) }}>
+                    <i className="fas fa-heart" aria-hidden="true"></i>
+                    <span className="srOnly">Click to open saved recipes</span>
+                </button>
 
-                            return (
-                                <li key={recipe.key} className="recipeListItem">
-                                    {/* <p>{recipe.name}</p> */}
-
-                                    <a href={recipe.recipeSource} target="_blank" rel="noreferrer">{recipe.name}</a>
-
-                                    <img src={recipe.foodImg} alt={`${recipe.foodName}`} onError={(event) => { event.target.src = notFoundPhoto; event.target.alt = "photo not found" }} />
-
-                                    <button onClick={() => { handleRemoveRecipe(recipe.key) }}>
-                                        <i className="fas fa-trash-alt" aria-hidden="true"></i>
-                                        <span className="srOnly">click to delete the recipe</span>
-                                    </button>
-                                </li>
-                            )
-                        })
-                    }
-                </ul>
+                <div className={showList ? "savedRecipeList toShow" : "savedRecipeList"}>
+                    <h3>Your Recipe List</h3>
+                    <p>{recipeListData.length} recipe{recipeListData.length > 1 ? "s" : ""} in list</p>
+                    <ul>
+                        {
+                            recipeListData.map((recipe) => {
+                                return (
+                                    <li key={recipe.key} className="recipeList">
+                                        <p>{recipe.name}</p>
+                                        <button onClick={() => { removeRecipeFunction(recipe.key) }}>
+                                            <i className="fas fa-trash-alt" aria-hidden="true"></i>
+                                            <span className="srOnly">click to delete the recipe</span>
+                                        </button>
+                                    </li>
+                                )
+                            })
+                        }
+                    </ul>
+                </div>
             </div>
         </>
     )
